Moved PEvent to PEventEmitter and PNetworkEvent to PNetworkEventEmitter
