+++ src/shared/models/PNetworkEventEmitter.ts
@@ -1,5 +1,5 @@
 import { PEventHandler } from '@shared/util/PapiUtil';
-import PEventEmitter from '@shared/models/PEvent';
+import PEventEmitter from '@shared/models/PEventEmitter';
 
 /**
  * Networked version of EventEmitter - accepts subscriptions to an event and runs the subscription callbacks when the event is emitted.
