+++ src/shared/services/papi.ts
@@ -8,7 +8,7 @@ import * as PapiUtil from '@shared/util/PapiUtil';
 // We need the WebViewService here to include on the papi, but WebViewService passes papi into WebViews
 // eslint-disable-next-line import/no-cycle
 import * as WebViewService from '@shared/services/WebViewService';
-import PEventEmitter from '@shared/models/PEvent';
+import PEventEmitter from '@shared/models/PEventEmitter';
 import logger from '@shared/util/logger';
 import { isRenderer } from '@shared/util/InternalUtil';
 import InternetService from '@shared/services/InternetService';
